Refactor JS code with objects and object constructors design pattern
The code is refactored to include use of object and object constructors to increase reusability and to reduce function coupling.
Changes in index.html and styles.css are only made to support JS code. Aesthetics in styles.css are not yet finished.
library.js is renamed to script.js, hence the deletion.
styles.scss and .gitignore are now tracked by git.